requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
apsw==3.36.0.post1
apturl==0.5.2
arrow==1.2.3
asgiref==3.7.2
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.2
attrs==21.2.0
auth0-python==4.4.0
backcall==0.2.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bidict==0.22.1
binaryornot==0.4.4
black==22.12.0
blinker==1.6.2
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chainlit==0.6.1
chardet==4.0.0
charset-normalizer==3.2.0
chromadb==0.3.26
click==8.1.5
clickhouse-connect==0.6.8
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
cookiecutter==2.1.1
cryptography==3.4.8
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.5.14
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
Deprecated==1.2.14
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
duckdb==0.8.1
duplicity==0.8.21
email-validator==2.0.0.post2
exceptiongroup==1.1.0
fastapi==0.97.0
fastapi-socketio==0.0.10
fasteners==0.14.1
feedparser==6.0.8
filetype==1.2.0
flake8==6.0.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.29.1
frozenlist==1.4.0
fs==2.4.12
future==0.18.2
gast==0.5.2
googleapis-common-protos==1.60.0
greenlet==1.1.2
grpcio==1.56.2
gyp==0.1
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httpcore==0.17.2
httplib2==0.20.2
httptools==0.5.0
httpx==0.24.1
humanfriendly==10.0
idna==3.3
ifaddr==0.1.7
importlib-metadata==6.8.0
iniconfig==2.0.0
ipython==7.31.1
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
jinja2-time==0.2.0
keyring==23.5.0
kiwisolver==1.3.2
langchain==0.0.250
langsmith==0.0.18
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
multidict==6.0.4
mypy-extensions==0.4.3
nest-asyncio==1.5.7
netifaces==0.11.0
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.2
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
orjson==3.9.1
overrides==7.3.1
packaging==21.3
pandas==2.0.3
paramiko==2.9.3
parso==0.8.1
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.6.2
pluggy==1.0.0
ply==3.11
posthog==3.0.1
prisma==0.9.1
prompt-toolkit==3.0.28
protobuf==4.23.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pulsar-client==3.2.0
py7zr==0.11.3+dfsg
pycairo==1.20.1
pychm==0.8.6
pycodestyle==2.10.0
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==1.10.9
pyflakes==3.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.3
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pytest==7.2.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-engineio==4.5.1
python-graphql-client==0.4.3
python-multipart==0.0.6
python-slugify==8.0.1
python-socketio==5.8.0
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2023.6.3
reportlab==3.6.8
repoze.lru==0.7
requests==2.31.0
requests-toolbelt==0.9.1
Routes==2.5.1
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.16
starlette==0.27.0
sympy==1.9
syncer==2.0.3
systemd-python==234
tenacity==8.2.2
text-unidecode==1.3
texttable==1.6.4
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.65.0
traitlets==5.1.1
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
unicodedata2==14.0.0
uptrace==1.19.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.22.0
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
xdg==5
xkit==0.0.0
yarl==1.9.2
zeroconf==0.38.3
zipp==1.0.0
zstandard==0.21.0
